+++ packages/kad-dht/test/peer-distance-list.spec.ts
@@ -6,17 +6,17 @@ import { PeerDistanceList } from '../src/peer-list/peer-distance-list.js'
 import * as kadUtils from '../src/utils.js'
 
 describe('PeerDistanceList', () => {
-  const p1 = peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE1')
-  const p2 = peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE2')
-  const p3 = peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE3')
-  const p4 = peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE4')
-  const p5 = peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE1')
-  const p6 = peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE5')
-  const p7 = peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE2')
+  const p1 = { id: peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE1'), multiaddrs: [] }
+  const p2 = { id: peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE2'), multiaddrs: [] }
+  const p3 = { id: peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE3'), multiaddrs: [] }
+  const p4 = { id: peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE4'), multiaddrs: [] }
+  const p5 = { id: peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE1'), multiaddrs: [] }
+  const p6 = { id: peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE5'), multiaddrs: [] }
+  const p7 = { id: peerIdFromString('12D3KooWSExt8hTzoaHEhn435BTK6BPNSY1LpTc1j2o9Gw53tXE2'), multiaddrs: [] }
 
   let key: Uint8Array
   before(async () => {
-    key = await kadUtils.convertPeerId(p1)
+    key = await kadUtils.convertPeerId(p1.id)
   })
 
   describe('basics', () => {
@@ -67,25 +67,25 @@ describe('PeerDistanceList', () => {
     })
 
     it('single closer peer', async () => {
-      const closer = await pdl.anyCloser([p6])
+      const closer = await pdl.anyCloser([p6.id])
 
       expect(closer).to.be.eql(true)
     })
 
     it('single further peer', async () => {
-      const closer = await pdl.anyCloser([p7])
+      const closer = await pdl.anyCloser([p7.id])
 
       expect(closer).to.be.eql(false)
     })
 
     it('closer and further peer', async () => {
-      const closer = await pdl.anyCloser([p6, p7])
+      const closer = await pdl.anyCloser([p6.id, p7.id])
 
       expect(closer).to.be.eql(true)
     })
 
     it('single peer equal to furthest in list', async () => {
-      const closer = await pdl.anyCloser([p2])
+      const closer = await pdl.anyCloser([p2.id])
 
       expect(closer).to.be.eql(false)
     })
@@ -98,7 +98,7 @@ describe('PeerDistanceList', () => {
 
     it('empty peer distance list', async () => {
       const pdl = new PeerDistanceList(key, 100)
-      const closer = await pdl.anyCloser([p1])
+      const closer = await pdl.anyCloser([p1.id])
 
       expect(closer).to.be.eql(true)
     })