+++ packages/peer-discovery-bootstrap/CHANGELOG.md
@@ -83,6 +83,15 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [10.0.23](https://github.com/libp2p/js-libp2p/compare/bootstrap-v10.0.22...bootstrap-v10.0.23) (2024-05-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^5.4.3 to ^5.4.4
+
 ## [10.0.22](https://github.com/libp2p/js-libp2p/compare/bootstrap-v10.0.21...bootstrap-v10.0.22) (2024-05-01)