+++ packages/protocol-identify/CHANGELOG.md
@@ -76,6 +76,18 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [2.0.1](https://github.com/libp2p/js-libp2p/compare/identify-v2.0.0...identify-v2.0.1) (2024-05-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^1.2.0 to ^1.2.1
+    * @libp2p/peer-record bumped from ^7.0.16 to ^7.0.17
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^5.4.3 to ^5.4.4
+
 ## [2.0.0](https://github.com/libp2p/js-libp2p/compare/identify-v1.0.21...identify-v2.0.0) (2024-05-01)